refactor(inventory): 将装备等级数据类型由byte改为short以支持更大范围，客户端最大显示255
- 修改Equip类中level和itemLevel字段类型为short
- 调整相关setter方法为short参数类型
- 更新gainEquip方法参数类型为short
- 修正ExpTable中超出等级经验的动态计算逻辑
- 修改相关DTO中level和itemLevel字段类型为short
- 修正InventoryService中读取和写入level及itemLevel的类型转换
- 更新ScrollHandler中记录装备原始等级变量为short
- 修改CharacterService中在线角色查询接口，增加多维度筛选参数和逻辑
- 优化查询结果，修复页面对象无stream方法问题，增加对在线时长和频道的准确过滤
- 维护离线角色查询逻辑，避免在线时长和频道筛选影响离线数据返回

+++ gms-server/src/main/java/org/gms/client/Character.java
@@ -9494,7 +9494,7 @@ public class Character extends AbstractCharacterObject {
      */
     public boolean gainEquip(int itemId, Short attStr, Short attDex, Short attInt, Short attLuk, Short attHp, Short attMp,
                              Short pAtk, Short mAtk, Short pDef, Short mDef, Short acc, Short avoid, Short hands, Short speed,
-                             Short jump, Byte upgradeSlot, Byte level, Byte itemLevel, Long expireTime,String owner, Short flag) {
+                             Short jump, Byte upgradeSlot, Short level, Short itemLevel, Long expireTime,String owner, Short flag) {
         Equip baseEquip = (Equip) ItemInformationProvider.getInstance().getEquipById(itemId);
         if (!ItemConstants.getInventoryType(itemId).equals(InventoryType.EQUIP) || baseEquip == null) {
             message(I18nUtil.getMessage("AbstractPlayerInteraction.gainEquip.message1"));
